친구 목록 컴포저블 분리
