+++ app/src/test/kotlin/team/ommaya/wequiz/android/HomeMainScreenSnapshot.kt
@@ -16,43 +16,31 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.takahirom.roborazzi.RobolectricDeviceQualifiers
 import com.github.takahirom.roborazzi.captureRoboImage
 import kotlinx.collections.immutable.persistentListOf
-import kotlinx.collections.immutable.toImmutableList
+import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
-import team.ommaya.wequiz.android.home.main.ExamNameAndIsWritingPair
-import team.ommaya.wequiz.android.home.main.HomeMainScreen
-import team.ommaya.wequiz.android.home.main.NicknameUuidScoreTriple
+import team.ommaya.wequiz.android.dummy.DummyExams
+import team.ommaya.wequiz.android.dummy.DummyFriendsRanking
+import team.ommaya.wequiz.android.home.HomeMainScreen
+import team.ommaya.wequiz.android.rule.SnapshotPathGeneratorRule
 
 @Config(qualifiers = RobolectricDeviceQualifiers.Pixel7Pro)
 @RunWith(AndroidJUnit4::class)
 class HomeMainScreenSnapshot {
+    @get:Rule
+    val snapshotPath = SnapshotPathGeneratorRule("HomeMainScreen")
+
     private val DummyProfileImage =
         Bitmap
             .createBitmap(100, 100, Bitmap.Config.ARGB_8888)
             .applyCanvas {
                 drawColor(Color.parseColor("#71F7FF"))
             }
-    private val DummyFriendsRanking =
-        List(5) { index ->
-            @Suppress("NAME_SHADOWING")
-            val index = index + 1
-            NicknameUuidScoreTriple("${index}번_친구_친구", index * 1_000, index * 100)
-        }.toImmutableList()
-    private val DummyExams =
-        List(5) { index ->
-            @Suppress("NAME_SHADOWING")
-            val index = index + 1
-            val boolean = index % 2 == 0
-            ExamNameAndIsWritingPair(
-                "${index}번${"_시험지".repeat(10)}".take(if (!boolean) 10 else 38),
-                boolean,
-            )
-        }.toImmutableList()
 
     @Test
     fun Empty() {
-        captureRoboImage("src/test/snapshots/HomeMainScreen/Empty.png") {
+        captureRoboImage(snapshotPath()) {
             HomeMainScreen(
                 nickname = "닉네임",
                 profileMessage = "",
@@ -65,7 +53,7 @@ class HomeMainScreenSnapshot {
 
     @Test
     fun OnlyExam() {
-        captureRoboImage("src/test/snapshots/HomeMainScreen/OnlyExam.png") {
+        captureRoboImage(snapshotPath()) {
             HomeMainScreen(
                 nickname = "닉네임",
                 profileMessage = "프로필 메시지",
@@ -78,7 +66,7 @@ class HomeMainScreenSnapshot {
 
     @Test
     fun FriendsRankingAndExam() {
-        captureRoboImage("src/test/snapshots/HomeMainScreen/FriendsRankingAndExam.png") {
+        captureRoboImage(snapshotPath()) {
             HomeMainScreen(
                 nickname = "닉네임",
                 profileMessage = "프로필 메시지\n프로필 메시지",
